+++ tests/query.rs
@@ -12,7 +12,7 @@ mod comparison {
     fn in_str_array() -> Result<(), String> {
         let item = Comparison::field("username").in_str_array(&["felix", "gerard"]);
         common::expect_assert_eq(
-            item.to_aql("i").as_str(),
+            item.aql_str("i").as_str(),
             r#"i.username IN ["felix", "gerard"]"#,
         )?;
         Ok(())
@@ -22,7 +22,7 @@ mod comparison {
     fn not_in_str_array() -> Result<(), String> {
         let item = Comparison::field("username").not_in_str_array(&["felix", "gerard"]);
         common::expect_assert_eq(
-            item.to_aql("i").as_str(),
+            item.aql_str("i").as_str(),
             r#"i.username NOT IN ["felix", "gerard"]"#,
         )?;
         Ok(())
@@ -31,10 +31,10 @@ mod comparison {
     #[test]
     fn in_array() -> Result<(), String> {
         let item = Comparison::field("age").in_array(&[13, 14, 15]);
-        common::expect_assert_eq(item.to_aql("i").as_str(), r#"i.age IN [13, 14, 15]"#)?;
+        common::expect_assert_eq(item.aql_str("i").as_str(), r#"i.age IN [13, 14, 15]"#)?;
         let item = Comparison::field("price").in_array(&[13.1, 14.5, 16.13]);
         common::expect_assert_eq(
-            item.to_aql("i").as_str(),
+            item.aql_str("i").as_str(),
             r#"i.price IN [13.1, 14.5, 16.13]"#,
         )?;
         Ok(())
@@ -43,10 +43,10 @@ mod comparison {
     #[test]
     fn not_in_array() -> Result<(), String> {
         let item = Comparison::field("age").not_in_array(&[13, 14, 15]);
-        common::expect_assert_eq(item.to_aql("i").as_str(), r#"i.age NOT IN [13, 14, 15]"#)?;
+        common::expect_assert_eq(item.aql_str("i").as_str(), r#"i.age NOT IN [13, 14, 15]"#)?;
         let item = Comparison::field("price").not_in_array(&[13.1, 14.5, 16.13]);
         common::expect_assert_eq(
-            item.to_aql("i").as_str(),
+            item.aql_str("i").as_str(),
             r#"i.price NOT IN [13.1, 14.5, 16.13]"#,
         )?;
         Ok(())
@@ -55,112 +55,112 @@ mod comparison {
     #[test]
     fn like() -> Result<(), String> {
         let item = Comparison::field("last_name").like("de %");
-        common::expect_assert_eq(item.to_aql("i").as_str(), r#"i.last_name LIKE "de %""#)?;
+        common::expect_assert_eq(item.aql_str("i").as_str(), r#"i.last_name LIKE "de %""#)?;
         Ok(())
     }
 
     #[test]
     fn not_like() -> Result<(), String> {
         let item = Comparison::field("last_name").not_like("de %");
-        common::expect_assert_eq(item.to_aql("i").as_str(), r#"i.last_name NOT LIKE "de %""#)?;
+        common::expect_assert_eq(item.aql_str("i").as_str(), r#"i.last_name NOT LIKE "de %""#)?;
         Ok(())
     }
 
     #[test]
     fn matches() -> Result<(), String> {
         let item = Comparison::field("last_name").matches(r#"^/[0.9]$"#);
-        common::expect_assert_eq(item.to_aql("i").as_str(), r#"i.last_name =~ "^/[0.9]$""#)?;
+        common::expect_assert_eq(item.aql_str("i").as_str(), r#"i.last_name =~ "^/[0.9]$""#)?;
         Ok(())
     }
 
     #[test]
     fn does_not_match() -> Result<(), String> {
         let item = Comparison::field("last_name").does_not_match(r#"^/[0.9]$"#);
-        common::expect_assert_eq(item.to_aql("i").as_str(), r#"i.last_name !~ "^/[0.9]$""#)?;
+        common::expect_assert_eq(item.aql_str("i").as_str(), r#"i.last_name !~ "^/[0.9]$""#)?;
         Ok(())
     }
 
     #[test]
     fn greater_than() -> Result<(), String> {
         let item = Comparison::field("age").greater_than(10);
-        common::expect_assert_eq(item.to_aql("i").as_str(), "i.age > 10")?;
+        common::expect_assert_eq(item.aql_str("i").as_str(), "i.age > 10")?;
         Ok(())
     }
 
     #[test]
     fn greater_or_equal() -> Result<(), String> {
         let item = Comparison::field("age").greater_or_equal(10);
-        common::expect_assert_eq(item.to_aql("i").as_str(), "i.age >= 10")?;
+        common::expect_assert_eq(item.aql_str("i").as_str(), "i.age >= 10")?;
         Ok(())
     }
 
     #[test]
     fn lesser_than() -> Result<(), String> {
         let item = Comparison::field("age").lesser_than(10);
-        common::expect_assert_eq(item.to_aql("i").as_str(), "i.age < 10")?;
+        common::expect_assert_eq(item.aql_str("i").as_str(), "i.age < 10")?;
         Ok(())
     }
 
     #[test]
     fn lesser_or_equal() -> Result<(), String> {
         let item = Comparison::field("age").lesser_or_equal(10);
-        common::expect_assert_eq(item.to_aql("i").as_str(), "i.age <= 10")?;
+        common::expect_assert_eq(item.aql_str("i").as_str(), "i.age <= 10")?;
         Ok(())
     }
 
     #[test]
     fn equals() -> Result<(), String> {
         let item = Comparison::field("age").equals(10);
-        common::expect_assert_eq(item.to_aql("i").as_str(), "i.age == 10")?;
+        common::expect_assert_eq(item.aql_str("i").as_str(), "i.age == 10")?;
         Ok(())
     }
 
     #[test]
     fn different_than() -> Result<(), String> {
         let item = Comparison::field("age").different_than(10);
-        common::expect_assert_eq(item.to_aql("i").as_str(), "i.age != 10")?;
+        common::expect_assert_eq(item.aql_str("i").as_str(), "i.age != 10")?;
         Ok(())
     }
 
     #[test]
     fn equals_str() -> Result<(), String> {
         let item = Comparison::field("name").equals_str("felix");
-        common::expect_assert_eq(item.to_aql("i").as_str(), r#"i.name == "felix""#)?;
+        common::expect_assert_eq(item.aql_str("i").as_str(), r#"i.name == "felix""#)?;
         Ok(())
     }
 
     #[test]
     fn different_than_str() -> Result<(), String> {
         let item = Comparison::field("name").different_than_str("felix");
-        common::expect_assert_eq(item.to_aql("i").as_str(), r#"i.name != "felix""#)?;
+        common::expect_assert_eq(item.aql_str("i").as_str(), r#"i.name != "felix""#)?;
         Ok(())
     }
 
     #[test]
-    fn is_null() -> Result<(), String> {
-        let item = Comparison::field("name").is_null();
-        common::expect_assert_eq(item.to_aql("i").as_str(), "i.name == null")?;
+    fn eq_null() -> Result<(), String> {
+        let item = Comparison::field("name").eq_null();
+        common::expect_assert_eq(item.aql_str("i").as_str(), "i.name == null")?;
         Ok(())
     }
 
     #[test]
     fn not_null() -> Result<(), String> {
         let item = Comparison::field("name").not_null();
-        common::expect_assert_eq(item.to_aql("i").as_str(), "i.name != null")?;
+        common::expect_assert_eq(item.aql_str("i").as_str(), "i.name != null")?;
         Ok(())
     }
 
     #[test]
-    fn is_true() -> Result<(), String> {
-        let item = Comparison::field("is_company").is_true();
-        common::expect_assert_eq(item.to_aql("i").as_str(), "i.is_company == true")?;
+    fn eq_true() -> Result<(), String> {
+        let item = Comparison::field("is_company").eq_true();
+        common::expect_assert_eq(item.aql_str("i").as_str(), "i.is_company == true")?;
         Ok(())
     }
 
     #[test]
-    fn is_false() -> Result<(), String> {
-        let item = Comparison::field("is_company").is_false();
-        common::expect_assert_eq(item.to_aql("i").as_str(), "i.is_company == false")?;
+    fn eq_false() -> Result<(), String> {
+        let item = Comparison::field("is_company").eq_false();
+        common::expect_assert_eq(item.aql_str("i").as_str(), "i.is_company == false")?;
         Ok(())
     }
 
@@ -170,27 +170,27 @@ mod comparison {
         #[test]
         fn all() -> Result<(), String> {
             let item = Comparison::all("emails").not_null();
-            common::expect_assert_eq(item.to_aql("i").as_str(), "i.emails ALL != null")?;
+            common::expect_assert_eq(item.aql_str("i").as_str(), "i.emails ALL != null")?;
             let item = compare!(all "emails").not_null();
-            common::expect_assert_eq(item.to_aql("i").as_str(), "i.emails ALL != null")?;
+            common::expect_assert_eq(item.aql_str("i").as_str(), "i.emails ALL != null")?;
             Ok(())
         }
 
         #[test]
         fn none() -> Result<(), String> {
-            let item = Comparison::none("emails").is_null();
-            common::expect_assert_eq(item.to_aql("i").as_str(), "i.emails NONE == null")?;
-            let item = compare!(none "emails").is_null();
-            common::expect_assert_eq(item.to_aql("i").as_str(), "i.emails NONE == null")?;
+            let item = Comparison::none("emails").eq_null();
+            common::expect_assert_eq(item.aql_str("i").as_str(), "i.emails NONE == null")?;
+            let item = compare!(none "emails").eq_null();
+            common::expect_assert_eq(item.aql_str("i").as_str(), "i.emails NONE == null")?;
             Ok(())
         }
 
         #[test]
         fn any() -> Result<(), String> {
-            let item = Comparison::any("authorizations").is_true();
-            common::expect_assert_eq(item.to_aql("i").as_str(), "i.authorizations ANY == true")?;
-            let item = compare!(any "authorizations").is_true();
-            common::expect_assert_eq(item.to_aql("i").as_str(), "i.authorizations ANY == true")?;
+            let item = Comparison::any("authorizations").eq_true();
+            common::expect_assert_eq(item.aql_str("i").as_str(), "i.authorizations ANY == true")?;
+            let item = compare!(any "authorizations").eq_true();
+            common::expect_assert_eq(item.aql_str("i").as_str(), "i.authorizations ANY == true")?;
             Ok(())
         }
     }
@@ -205,7 +205,7 @@ mod filter {
     fn provides_correct_string() -> Result<(), String> {
         let filter = Filter::new(Comparison::field("username").equals_str("felix"))
             .and(Comparison::field("age").greater_than(15));
-        let filter_str = filter.to_aql("i");
+        let filter_str = filter.aql_str("i");
         common::expect_assert_eq(
             filter_str.as_str(),
             r#"i.username == "felix" && i.age > 15"#,
@@ -219,7 +219,7 @@ mod filter {
             .and(Comparison::field("company_age").greater_than(15))
             .or(Comparison::any("emails").like("%gmail.com"))
             .and(Comparison::field("roles").in_str_array(&["SHIPPER", "FORWARDER"]));
-        let filter_str = filter.to_aql("i");
+        let filter_str = filter.aql_str("i");
         common::expect_assert_eq(
             filter_str.as_str(),
             "\
@@ -252,7 +252,7 @@ mod query {
                         .prune(Comparison::statement("1").equals(1).into()),
                 );
             common::expect_assert_eq(
-                query.to_aql().as_str(),
+                query.aql_str().as_str(),
                 "\
             FOR b in Companies \
                 FILTER b.emails ANY LIKE \"%gmail.com\" \
@@ -290,7 +290,7 @@ mod query {
                         ),
                 );
             common::expect_assert_eq(
-                query.to_aql().as_str(),
+                query.aql_str().as_str(),
                 "\
             FOR d in Companies \
                 FILTER d.emails ANY LIKE \"%gmail.com\" \
@@ -323,7 +323,7 @@ mod query {
                         .prune(Comparison::statement("1").equals(1).into()),
                 );
             common::expect_assert_eq(
-                query.to_aql().as_str(),
+                query.aql_str().as_str(),
                 "\
             FOR b in Companies \
                 FILTER b.emails ANY LIKE \"%gmail.com\" \
@@ -361,7 +361,7 @@ mod query {
                         ),
                 );
             common::expect_assert_eq(
-                query.to_aql().as_str(),
+                query.aql_str().as_str(),
                 "\
             FOR d in Companies \
                 FILTER d.emails ANY LIKE \"%gmail.com\" \
@@ -386,7 +386,7 @@ mod query {
             .limit(5, None)
             .distinct();
         common::expect_assert_eq(
-            query.to_aql().as_str(),
+            query.aql_str().as_str(),
             "\
         FOR a in Companies \
             FILTER a.emails ANY LIKE \"%gmail.com\" \
@@ -406,7 +406,7 @@ mod query {
             .limit(5, None)
             .distinct();
         common::expect_assert_eq(
-            query.to_aql().as_str(),
+            query.aql_str().as_str(),
             "FOR a in Companies \
                         FILTER a.emails ANY LIKE \"%gmail.com\" \
                         SORT a.company_name ASC, a.company_age DESC \
@@ -419,12 +419,12 @@ mod query {
     #[test]
     fn order_of_operations_works() -> Result<(), String> {
         let query = Query::new("Users")
-            .filter(Comparison::field("active").is_true().into())
+            .filter(Comparison::field("active").eq_true().into())
             .sort("age", None)
             .limit(5, None)
             .filter(Comparison::field("gender").equals_str("f").into());
         common::expect_assert_eq(
-            query.to_aql().as_str(),
+            query.aql_str().as_str(),
             "FOR a in Users \
                     FILTER a.active == true \
                     SORT a.age ASC \
@@ -448,7 +448,7 @@ mod query {
             .limit(5, None)
             .distinct();
         common::expect_assert_eq(
-            query.to_aql().as_str(),
+            query.aql_str().as_str(),
             "FOR a in Companies \
                        FILTER a.emails ANY LIKE \"%gmail.com\" && a.id > 10 \
                        SORT a.company_name DESC, a.company_age ASC \
@@ -461,7 +461,7 @@ mod query {
     #[test]
     fn empty_query_works() -> Result<(), String> {
         let query = Query::new("Companies");
-        common::expect_assert_eq(query.to_aql().as_str(), "FOR a in Companies return a")?;
+        common::expect_assert_eq(query.aql_str().as_str(), "FOR a in Companies return a")?;
         Ok(())
     }
 }
@@ -688,11 +688,7 @@ mod call {
         factory(&connection);
         let query = Query::new("Dish");
         let cursor = query.raw_call_in_batches(&connection, 1).unwrap();
-        let mut count = 0;
-        for _result in cursor.into_iter() {
-            count += 1;
-        }
-        common::expect_assert_eq(count, 5)?;
+        common::expect_assert_eq(cursor.into_iter().count(), 5)?;
         Ok(())
     }